% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_data.R
\name{raw_data}
\alias{raw_data}
\title{Get paths to \code{sqlr.per3} raw data}
\usage{
raw_data(type = NULL, file = NULL)
}
\arguments{
\item{type}{(optional) a string indicating the file type of the raw dataset
(default: \code{NULL}).}

\item{file}{(optional) a \code{character} object indicating the file name(s) of
the raw dataset.}
}
\value{
\itemize{
\item If \code{type = NULL}, a \code{character} object with all file type names
available.
\item If \code{type != NULL} and \code{file = NULL}, a \code{character} object with all file
names available from type.
\item If \code{type != NULL} and \code{file != NULL}, a string with the file name path.
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{raw_data()} returns the raw data paths of the \code{sqlr.per3} package.
}
\examples{
\dontrun{
raw_data()
}
}
\concept{utility functions}
