% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlr.per3-package.R
\docType{package}
\name{sqlr.per3-package}
\alias{sqlr.per3}
\alias{sqlr.per3-package}
\title{sqlr.per3: R Package for the PER3 Gene and Circadian Rhythms Project}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

R package for the project: PER3 gene and circadian rhythms: 
      a systematic quantitative literature review.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gipsousp.github.io/sqlr.per3/}
  \item \url{https://github.com/gipsousp/sqlr.per3/}
  \item Report bugs at \url{https://github.com/gipsousp/sqlr.per3/issues/}
}

}
\author{
\strong{Maintainer}: Daniel Vartanian \email{danvartan@gmail.com} (\href{https://orcid.org/0000-0001-7782-759X}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Juliana Viana Mendes \email{mendes.julianav@gmail.com} (\href{https://orcid.org/0000-0002-8222-5240}{ORCID}) [copyright holder]
  \item Ana Amelia Benedito-Silva \email{aamelia@usp.br} (\href{https://orcid.org/0000-0003-4976-2623}{ORCID}) [scientific advisor]
  \item Mario Pedrazzoli \email{pedrazzo@usp.br} (\href{https://orcid.org/0000-0002-5257-591X}{ORCID}) [scientific advisor]
}

Other contributors:
\itemize{
  \item Interdisciplinary Sleep Research Group (GIPSO) [funder, copyright holder]
  \item University of Sao Paulo (USP) [funder]
}

}
\keyword{internal}
